MobileNet alpha update
Load alpha on login screen so it's ready to use when loading the model
